wip: create styles module
This is a WIP that didn't work out so well. It would be really nice if
stylix let you configure theme settings on a per-application basis, but
it doesn't. The issue is that the theme you want in your terminal for
optimal syntax highlighting isn't always the theme you want for Gnome.
The same thing goes for fonts and font sizes.